third_party/WebKit/LayoutTests/fast/css-grid-layout/floated-grid-container-percentage-tracks.html

<!DOCTYPE html>
<link href="resources/grid.css" rel="stylesheet">
<style>

.wrapper {
    position: relative;
    clear: both;
}

.grid {
    float: left;
    font: 10px/1 Ahem;
}

.fixedSize {
    width: 200px;
    height: 200px;
}

.oneRowOneColumn {
    grid-template-rows: 100%;
    grid-template-columns: 100%;
}

.twoRowsTwoColumns {
    grid-template-rows: 25% 75%;
    grid-template-columns: 60% 40%;
}

</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div id="log"></div>

<p>This test checks that percentage tracks are properly resolved for floated grid containers.</p>

<div class="wrapper">
    <div class="grid oneRowOneColumn">
        <div class="firstRowFirstColumn"
            data-offset-x="0" data-offset-y="0" data-expected-width="30" data-expected-height="10">
            XXX
        </div>
    </div>
</div>

<div class="wrapper">
    <div class="grid oneRowOneColumn fixedSize">
        <div class="firstRowFirstColumn"
            data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="200">
            XXX
        </div>
    </div>
</div>

<div class="wrapper">
    <div class="grid twoRowsTwoColumns">
        <div class="firstRowFirstColumn"
            data-offset-x="0" data-offset-y="0" data-expected-width="30" data-expected-height="10">
            XXX
        </div>
        <div class="firstRowSecondColumn"
            data-offset-x="30" data-offset-y="0" data-expected-width="30" data-expected-height="10">
            XXX
        </div>
        <div class="secondRowFirstColumn"
            data-offset-x="0" data-offset-y="10" data-expected-width="30" data-expected-height="10">
            XXX
        </div>
        <div class="secondRowSecondColumn"
            data-offset-x="30" data-offset-y="10" data-expected-width="30" data-expected-height="10">
            XXX
        </div>
    </div>
</div>

<div class="wrapper">
    <div class="grid twoRowsTwoColumns fixedSize">
        <div class="firstRowFirstColumn"
            data-offset-x="0" data-offset-y="0" data-expected-width="120" data-expected-height="50">
            XXX
        </div>
        <div class="firstRowSecondColumn"
            data-offset-x="120" data-offset-y="0" data-expected-width="80" data-expected-height="50">
            XXX
        </div>
        <div class="secondRowFirstColumn"
            data-offset-x="0" data-offset-y="50" data-expected-width="120" data-expected-height="150">
            XXX
        </div>
        <div class="secondRowSecondColumn"
            data-offset-x="120" data-offset-y="50" data-expected-width="80" data-expected-height="150">
            XXX
        </div>
    </div>
</div>

</body>
